Add unit/acceptance tests and update README
This adds working unit tests and an acceptance test in beaker to test
the dynect provider.  The README is also updated with examples and
better descriptions of about everything.

After this is merged, I'll be able to look into changing the dynect
provider more easily to use the (new) fog gem for dynect instead of the
manual hacks I use to only implement a subset of DNS records.